ml_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
numpy==2.1.2
sentence-transformers==3.1.1
faiss-cpu==1.13.0
openai==1.51.0

